GetGitRepo now includes OpenTerminalAtFolder functionality
GetGitRepo loads OpenTerminalAtFolder so that both functions can be bound to the same keyboard shortcut.

If the Finder is the active application, the script will open Terminal at the selected folder and also open the related repo in Chrome (if Chrome is open).

If Finder is NOT the frontmost application, the script will try to fork and clone the git repo just as before (provided a repo is available).